Stop launch of customer app to prevent crash
The first log in of the venue app generates the database.  If the customer app is launched before that first login on venue app, then the customer app will crash.  Better to disable automatic launch and aske the user to do it, rather than risk user crashing customer app straightaway.